requirements.txt
GitPython==3.1.31
python-gitlab==3.15.0